.prettier contracts

+++ packages/contracts/contracts/CollSurplusPool.sol
@@ -8,7 +8,6 @@ import "./Dependencies/CheckContract.sol";
 import "./Dependencies/console.sol";
 import "./CollSurplusPoolStorage.sol";
 
-
 contract CollSurplusPool is CollSurplusPoolStorage, CheckContract, ICollSurplusPool {
     using SafeMath for uint256;
     // --- Events ---
@@ -17,20 +16,16 @@ contract CollSurplusPool is CollSurplusPoolStorage, CheckContract, ICollSurplusP
     event TroveManagerAddressChanged(address _newTroveManagerAddress);
     event ActivePoolAddressChanged(address _newActivePoolAddress);
 
-    event CollBalanceUpdated(address indexed _account, uint _newBalance);
-    event EtherSent(address _to, uint _amount);
-    
+    event CollBalanceUpdated(address indexed _account, uint256 _newBalance);
+    event EtherSent(address _to, uint256 _amount);
+
     // --- Contract setters ---
 
     function setAddresses(
         address _borrowerOperationsAddress,
         address _troveManagerAddress,
         address _activePoolAddress
-    )
-        external
-        override
-        onlyOwner
-    {
+    ) external override onlyOwner {
         checkContract(_borrowerOperationsAddress);
         checkContract(_troveManagerAddress);
         checkContract(_activePoolAddress);
@@ -42,26 +37,24 @@ contract CollSurplusPool is CollSurplusPoolStorage, CheckContract, ICollSurplusP
         emit BorrowerOperationsAddressChanged(_borrowerOperationsAddress);
         emit TroveManagerAddressChanged(_troveManagerAddress);
         emit ActivePoolAddressChanged(_activePoolAddress);
-
-        
     }
 
     /** Returns the ETH state variable at ActivePool address.
        Not necessarily equal to the raw ether balance - ether can be forcibly sent to contracts. */
-    function getETH() external view override returns (uint) {
+    function getETH() external view override returns (uint256) {
         return ETH;
     }
 
-    function getCollateral(address _account) external view override returns (uint) {
+    function getCollateral(address _account) external view override returns (uint256) {
         return balances[_account];
     }
 
     // --- Pool functionality ---
 
-    function accountSurplus(address _account, uint _amount) external override {
+    function accountSurplus(address _account, uint256 _amount) external override {
         _requireCallerIsTroveManager();
 
-        uint newAmount = balances[_account].add(_amount);
+        uint256 newAmount = balances[_account].add(_amount);
         balances[_account] = newAmount;
 
         emit CollBalanceUpdated(_account, newAmount);
@@ -69,7 +62,7 @@ contract CollSurplusPool is CollSurplusPoolStorage, CheckContract, ICollSurplusP
 
     function claimColl(address _account) external override {
         _requireCallerIsBorrowerOperations();
-        uint claimableColl = balances[_account];
+        uint256 claimableColl = balances[_account];
         require(claimableColl > 0, "CollSurplusPool: No collateral available to claim");
 
         balances[_account] = 0;
@@ -87,19 +80,16 @@ contract CollSurplusPool is CollSurplusPoolStorage, CheckContract, ICollSurplusP
     function _requireCallerIsBorrowerOperations() internal view {
         require(
             msg.sender == borrowerOperationsAddress,
-            "CollSurplusPool: Caller is not Borrower Operations");
+            "CollSurplusPool: Caller is not Borrower Operations"
+        );
     }
 
     function _requireCallerIsTroveManager() internal view {
-        require(
-            msg.sender == troveManagerAddress,
-            "CollSurplusPool: Caller is not TroveManager");
+        require(msg.sender == troveManagerAddress, "CollSurplusPool: Caller is not TroveManager");
     }
 
     function _requireCallerIsActivePool() internal view {
-        require(
-            msg.sender == activePoolAddress,
-            "CollSurplusPool: Caller is not Active Pool");
+        require(msg.sender == activePoolAddress, "CollSurplusPool: Caller is not Active Pool");
     }
 
     // --- Fallback function ---
